Refact: Created separate components for Admin Drawer
