+++ Frontend/src/Component/Admin/Drawer/drawer.module.css
@@ -1,0 +1,4 @@
+.navbarLogo {
+  width: 140px;
+  cursor: pointer;
+}
