--- Frontend/src/pages/admin/admin.module.css
@@ -1,5 +1,0 @@
-.navbarLogo {
-  width: 120px;
-  cursor: pointer;
-
-}
